// frontend/src/company.d.ts
export interface Medication {
  MedicationID: number;
  Name: string;
  UsagePerDay: number;
  Dosage: number;
  Effects: string;
  SideEffects: string | null;
}